+++ server/db/models/product.js
@@ -11,10 +11,10 @@ const Product = db.define('product', {
     allowNull: false
   },
   price: {
-    type: Sequelize.FLOAT,
+    type: Sequelize.INTEGER,
     allowNull: false,
     validate: {
-      min: 0.0
+      min: 0
     }
   },
   description: {
